+++ src/main/java/com/project/serverapp/repo/RegionRepo.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RegionRepo extends JpaRepository<Region, Long> {}
+public interface RegionRepo extends JpaRepository<Region, Integer> {}
